fixed a few typos in the api
git-svn-id: https://commonmusic.svn.sourceforge.net/svnroot/commonmusic/trunk@1463 d60aafaf-7936-0410-ba4d-d307febf7868

+++ grace/src/Toolbox.cpp
@@ -74,7 +74,7 @@ float Toolbox::keynum_to_hertz (float kn) {
   return A00 * pow(2, ((kn + 3) / 12));
 }
 
-float Toolbox::keynum_to_pc (float kn) {
+int Toolbox::keynum_to_pc (float kn) {
   return ((int)kn) % 12;
 }
